fix: Plus icon works now
The plus icon in the homepage works and the icon is rendered correctly. If you now click on it, a form that was automatically generated with the generation of the scaffold will be rendered and even an image can be added in the form. The car will be then rendered on the homepage.