feat: Functioning like button added
A functioning like button was added so that posts can now be liked. For that, a new method was implemented in the CarsController that's named 'like' which increments the likes attribute of the post that is liked by 1. Posts can be liked infinitely by anyone though. This will be fixed soon.