fix: Like function now working properly
The like button for the posts should now work properly. If you want to like a post, you need to be logged in otherwise it will redirect you to the login page. When a post is liked the icon should change from an empty heart icon to a filled out heart icon.