add: New page "About"
A new page about is now available. The purpose of this page is to provide more information about the project, how it's build, the dependencies, contributors and so on. This page still has some way to go until it's finished but it is already available. Additionally, a new edit function was added for posts so you can now edit posts if you are the owner.